fix(next-intl): move provider out of root layout
