+++ src/app/components/ClientProviders/ClientProviders.tsx
@@ -2,75 +2,27 @@
 
 import React, { ComponentProps } from 'react';
 import { CookiesProvider } from 'next-client-cookies';
-import { useSocket } from '@/lib/socket/useSocket';
-import { useAction } from 'next-safe-action/hook';
-import { revalidateAppAction } from '@/actions/app-actions/revalidate-app';
-import { useTranslations } from 'next-intl';
-import toast from 'react-hot-toast';
+import { AbstractIntlMessages, NextIntlClientProvider } from 'next-intl';
 import { ThemeProvider } from './ThemeProvider';
+import { SocketProvider } from './SocketProvider/SocketProvider';
 
 type Props = {
   children: React.ReactNode;
   cookies: ComponentProps<typeof CookiesProvider>['value'];
   initialTheme?: string;
+  locale?: string;
+  messages: AbstractIntlMessages;
 };
 
-export const ClientProviders = ({ children, initialTheme, cookies }: Props) => {
-  const revalidateAppMutation = useAction(revalidateAppAction);
-  const t = useTranslations();
-
-  useSocket({
-    onEvent: (event, data) => {
-      revalidateAppMutation.execute({ id: data.appId });
-
-      switch (event) {
-        case 'install_success':
-          toast.success(t('APP_INSTALL_SUCCESS', { id: data.appId }));
-          break;
-        case 'install_error':
-          toast.error(t('APP_ERROR_APP_FAILED_TO_INSTALL', { id: data.appId }));
-          break;
-        case 'start_success':
-          toast.success(t('APP_START_SUCCESS', { id: data.appId }));
-          break;
-        case 'start_error':
-          toast.error(t('APP_ERROR_APP_FAILED_TO_START', { id: data.appId }));
-          break;
-        case 'stop_success':
-          toast.success(t('APP_STOP_SUCCESS', { id: data.appId }));
-          break;
-        case 'stop_error':
-          toast.error(t('APP_ERROR_APP_FAILED_TO_STOP', { id: data.appId }));
-          break;
-        case 'uninstall_success':
-          toast.success(t('APP_UNINSTALL_SUCCESS', { id: data.appId }));
-          break;
-        case 'uninstall_error':
-          toast.error(t('APP_ERROR_APP_FAILED_TO_UNINSTALL', { id: data.appId }));
-          break;
-        case 'update_success':
-          toast.success(t('APP_UPDATE_SUCCESS', { id: data.appId }));
-          break;
-        case 'update_error':
-          toast.error(t('APP_ERROR_APP_FAILED_TO_UPDATE', { id: data.appId }));
-          break;
-        case 'reset_success':
-          toast.success(t('APP_RESET_SUCCESS', { id: data.appId }));
-          break;
-        case 'reset_error':
-          toast.error(t('APP_ERROR_APP_FAILED_TO_RESET', { id: data.appId }));
-          break;
-        default:
-          break;
-      }
-    },
-    selector: { type: 'app' },
-  });
-
+export const ClientProviders = ({ children, initialTheme, cookies, locale, messages }: Props) => {
   return (
-    <CookiesProvider value={cookies}>
-      <ThemeProvider initialTheme={initialTheme}>{children}</ThemeProvider>
-    </CookiesProvider>
+    <NextIntlClientProvider locale={locale} messages={messages} timeZone={Intl.DateTimeFormat().resolvedOptions().timeZone}>
+      <SocketProvider>
+        <CookiesProvider value={cookies}>
+          <ThemeProvider initialTheme={initialTheme}>{children}</ThemeProvider>
+        </CookiesProvider>
+      </SocketProvider>
+    </NextIntlClientProvider>
   );
 };
